code refactor for harnesses
