+++ src/harness/__init__.py
@@ -13,13 +13,9 @@ Key exports:
 
 from .agent import Agent, AgentTrace, OptionsProvider
 from .sdk_config import set_sdk, get_sdk, is_claude_sdk, is_opencode_sdk
-from .options_utils import (
-    build_options,
-    build_claudecode_options,
-    build_opencode_options,
-    resolve_project_root,
-    resolve_data_dirs,
-)
+from .utils import build_options, resolve_project_root, resolve_data_dirs
+from .claude.options import build_claudecode_options
+from .opencode.options import build_opencode_options
 
 __all__ = [
     "Agent",
