+++ src/harness/claude/__init__.py
@@ -1,0 +1,6 @@
+"""Claude SDK harness — option building and execution."""
+
+from .options import build_claudecode_options
+from .executor import execute_query, parse_response
+
+__all__ = ["build_claudecode_options", "execute_query", "parse_response"]
