+++ src/harness/claude/executor.py
