+++ src/harness/opencode/executor.py
